Extract the cause-related code to a separate module (as is)
